+++ source/client/utils/Helpers.ts
@@ -44,124 +44,3 @@ export function getMeshTransform(root : Object3D, current: Object3D)
 
     return result;
 }
-
-export function injectVertexShaderCode(shader: string) : string {
-    shader = '// Zone map support\n \
-        #if defined(USE_ZONEMAP)\n \
-            varying vec2 vZoneUv;\n \
-        #endif\n \
-        \n \
-        #ifdef MODE_XRAY\n \
-            varying float vIntensity;\n \
-        #endif\n \
-        \n'.concat(shader);
-
-    shader = shader.slice(0,shader.lastIndexOf('}')).concat(
-        '\n \
-        #ifdef MODE_NORMALS\n \
-            vNormal = normal;\n \
-        #endif\n \
-        \n \
-        #ifdef MODE_XRAY\n \
-            vIntensity = pow(abs(1.0 - abs(dot(vNormal, vec3(0.0, 0.0, 1.0)))), 3.0);\n \
-        #endif\n \
-        \n \
-        // Zone map support\n \
-        #if defined(USE_ZONEMAP)\n \
-            #if defined(USE_MAP)\n \
-                vZoneUv = (mapTransform * vec3(vMapUv, 1)).xy;\n \
-            #else\n \
-                vZoneUv = uv;\n \
-            #endif\n \
-        #endif\n \
-        }'
-    )
-
-    return shader;
-}
-
-export function injectFragmentShaderCode(shader: string) {
-    shader = shader.replace(
-        'void main() {',
-        '// Zone map support\n \
-        #if defined(USE_ZONEMAP)\n \
-            varying vec2 vZoneUv;\n \
-            uniform sampler2D zoneMap;\n \
-        #endif\n \
-        \n \
-        #ifdef MODE_XRAY\n \
-            varying float vIntensity;\n \
-        #endif\n \
-        \n \
-        #ifdef CUT_PLANE\n \
-            uniform vec3 cutPlaneColor;\n \
-        #endif\n \
-        \n \
-        void main() {'
-    )
-
-    shader = shader.replace(
-        '#include <opaque_fragment>',
-        '#include <opaque_fragment>\n \
-        \n \
-        #ifdef USE_ZONEMAP\n \
-            vec4 zoneColor = texture2D(zoneMap, vZoneUv);\n \
-        \n \
-            #ifdef OVERLAY_ALPHA\n \
-                gl_FragColor += mix(vec4(0.0, 0.0, 0.0, 1.0), vec4(zoneColor.rgb, 1.0), zoneColor.a);\n \
-            #endif\n \
-            #ifndef OVERLAY_ALPHA\n \
-                gl_FragColor = mix(gl_FragColor, vec4(zoneColor.rgb, 1.0), zoneColor.a);\n \
-            #endif\n \
-        #endif\n \
-        \n'
-    )
-
-    shader = shader.slice(0,shader.lastIndexOf('}')).concat(
-        '\n \
-        #ifdef CUT_PLANE\n \
-        if (!gl_FrontFacing) {\n \
-            gl_FragColor = vec4(cutPlaneColor.rgb, 1.0);\n \
-        }\n \
-        #endif\n \
-        \n \
-        #ifdef MODE_NORMALS\n \
-            gl_FragColor = vec4(vec3(normal * 0.5 + 0.5), 1.0);\n \
-        #endif\n \
-        \n \
-        #ifdef MODE_XRAY\n \
-            gl_FragColor = vec4(vec3(0.4, 0.7, 1.0) * vIntensity, 1.0);\n \
-        #endif\n \
-        }'
-    )
-
-    return shader;
-}
-
-export function addCustomMaterialDefines(material: Material) {
-    material.defines ??= {};
-
-    material.defines["OBJECTSPACE_NORMALMAP"] = false;
-    material.defines["MODE_NORMALS"] = false;
-    material.defines["MODE_XRAY"] = false;
-    material.defines["CUT_PLANE"] = false;
-    material.defines["USE_ZONEMAP"] = false;
-    material.defines["OVERLAY_ALPHA"] = false;
-}
-
-export function extendShaders(material: Material) {
-    const uniforms = {
-        cutPlaneColor: { value: _vec3 },
-        zoneMap: { value: null }
-    };
-
-    material.onBeforeCompile = (shader) => {
-        shader.vertexShader = injectVertexShaderCode(shader.vertexShader);
-        shader.fragmentShader = injectFragmentShaderCode(shader.fragmentShader);
-
-        // add custom uniforms
-        shader.uniforms.cutPlaneColor = uniforms.cutPlaneColor;
-        shader.uniforms.zoneMap = uniforms.zoneMap;
-        material.userData.shader = shader;
-    }
-}
